perf: separate test files for performance
